backend/requirements.txt
django
django-extensions
cryptography
pyhanko
reportlab
pypdf
mysqlclient
python-dotenv